+++ changelog.md
@@ -1,8 +1,8 @@
 ### Additions
-* Added `canApplyKey` method to `LootablesApi`. This method will check if the player provided can apply the given key, before attempting to roll loot.
+* Added new `Loader` to `LootableItem.LootableData` for compositional loading as needed.
 
 ### Changes
-* Elevated the `CODEC` and `PACKET_CODEC` of `LootableItem.LootableData` to public for use in codec construction outside of Lootables.
+* None.
 
 ### Fixes
 * None.